Adding comments and polishing the code

+++ game/src/Bullet.h
@@ -11,17 +11,17 @@ public:
 		timeToLive = 1.5;
 		bulletTexture = theBulletTexture;
 		direction = theDirection;
-		scale = 0.1;
-		rotation = 0;
-		pivot.x = bulletTexture.width / 2 * scale;
-		pivot.y = bulletTexture.height / 2 * scale;
+		scale = 0.1f;
+		rotation = 0.0f;
+		pivot.x = bulletTexture.width / 2.0f * scale;
+		pivot.y = bulletTexture.height / 2.0f * scale;
 		objectRadius = pivot.x;
 		position = thePosition;
 		position.x = position.x - (2 + theDirection.x) * pivot.x;
 		position.y = position.y - (2 + theDirection.y) * pivot.y;
-		rotation = atan2(direction.y, direction.x) * RAD2DEG + 90;
-		objectRect.width = bulletTexture.width;
-		objectRect.height = bulletTexture.height;
+		rotation = static_cast<float>(atan2(direction.y, direction.x) * static_cast<float>(RAD2DEG) + 90.0f);
+		objectRect.width = static_cast<float>(bulletTexture.width);
+		objectRect.height = static_cast<float>(bulletTexture.height);
 		speed = 700;
 		ImCircular = false;
 	}
