+++ game/src/Globals.h
@@ -9,7 +9,7 @@
 // Declatation of global variables
 int screenWidth = 1280;
 int screenHeight = 720;
-Rectangle screen{ 0, 0, screenWidth, screenHeight };
+Rectangle screen{ 0.0f, 0.0f, static_cast<float>(screenWidth), static_cast<float>(screenHeight) };
 int currentPoints = 0;
 int highScore = 0;
 Texture2D bulletTexture;
